@@ -17,10 +17,10 @@
 	<description>Demo project for Spring Boot</description>
 
 	<properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 		<lombok.version>1.18.34</lombok.version>
 		<mysql.connector.version>8.0.33</mysql.connector.version>
-		<maven.compiler.plugin.version>3.10.1</maven.compiler.plugin.version>
+		<maven.compiler.plugin.version>3.11.0</maven.compiler.plugin.version>
 	</properties>
 
 	<dependencies>
