@@ -1,3 +1,6 @@
+// ► Modelo de solicitud para crear o actualizar usuarios
+//   Esta clase representa los datos que el cliente envía en el body.
+//   Incluye validación para campos obligatorios como el nombre.
 package com.proyec.crud.model;
 
 import jakarta.validation.constraints.NotBlank;
@@ -12,6 +15,5 @@ public class UserRq {
     @NotBlank(message = "Este campo es obligatorio")
     private String name;
     private String email;
-    private  String password;
-
+    private String password;
 }
