@@ -1,3 +1,6 @@
+// ► Modelo de respuesta del usuario
+//   Clase usada para devolver datos del usuario al cliente sin exponer información sensible.
+
 package com.proyec.crud.model;
 
 import lombok.AllArgsConstructor;
@@ -7,9 +10,8 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserRs  {
+public class UserRs {
     private String id;
     private String name;
     private String email;
-
 }
